bugfixing loot analyzer
hopefully the loot analyzer won't mark a stash as "visited" when we walk
on it on our way to eg. a monster (using "m<dir>"). currently the loot
analyzer should set command to LOOK with priority PRIORITY_LOOK so
she'll pick up thrown weapons while heading for the monster, although we
could reduce the priority if we don't want her to do that.
one change is that weapons she throws while blind and she then walks
over won't be looked at, unlike previously. this is probably a good
thing in case we kill a *trice while blind as it's less likely that
we'll "feel what's on the ground" and die