made best_priority public and sorting analyzers
saiph now got a global variable "best_priority", this can be used by
analyzers to determine if there's any point at all for them to do their
stuff. for example, if we're fighting with priority 450, then it's
pointless for explore to path all over the map, because we know that the
fight analyzer will end up with a higher priority anyways. but there's
an issue here, what if we call explore->analyze() before
fight->analyze() (since we've ordered the analyzers alphabetically this
is also the case)? then we'll still do all the explore stuff in explore
because best_priority haven't been set to 450 by fight->analyze() yet.
this is where sorting analyzers comes into the picture:
we're now sorting analyzers each time best_analyzers is not the first
analyzer in the vector. however, as i'm writing this i realize that it
got too late when i implemented this, because the implementation is
poor:
1. when one analyzer request another analyzer to do something (for
   example, Health analyzer request that the Potion analyzer quaff a
   potion, then Health analyzer will end up with priority -1 and Potion
   with priority 500 (or something). in this case, sorting the vector
   makes no sense at all.
2. priorities may vary a lot, sorting the vector all the time is just
   time consuming. a better solution is if every analyzer set an
   "expected complexity" and we sort the analyzers once after that. a
   "complex" method being one that demands a lot of resources (for
   example, both the explore and fight analyzers are "complex" as they
   do pathfinding, health is not very complex as it only checks some
   stats and elbereth got almost no complexity at all since it mostly
   just respond to requests and parse some messages).
   complexity also have to be weighted up with "how likely is this
   method to be called?", the explore analyzers is very active, and it's
   probably a good idea to not set this "expected complexity" too high
   as explore wil have the highest priority quite often.
   i'm thinking of just making a small bunch of defines like
   "COMPLEXITY_VERY_HIGH", "COMPLEXITY_HIGH", and so on