call previous analyzer first on question/menu
this is in a way making a "multi command" "atomic". eg. if door analyzer
sends command "a" to apply a key, then the next update() door will be
called first, which will set priority to PRIORITY_CONTINUE_ACTION.
unihorn may also respond to the question about what to apply, but that
too will set the priority to PRIORITY_CONTINUE_ACTION, and since that's
not greater than PRIORITY_CONTINUE_ACTION, the door analyzer will "win".
in a way, this is slightly dodgy, and ofcourse, if we're using different
priorities it will fail, but *every question shall be answered with
priority PRIORITY_CONTINUE_ACTION* so if you do it right, it should not
be a problem.
credit to doy for suggestions & tips on how this is done in taeb.