simplified Genocide
some might react "but now we might geno 'h' even if both flayers already
are genoed!", and this is quite correct. however, at the current time
it's unlikely that we get that many genoes, and even in the long run we
can't have a huge bunch of booleans to determine whether we should geno
a class when we've genoed the nasties in that class.
for the moment, this will work well enough.