no longer using pick-axe as weapon
pick-axe is now only picked up as a tool. she gets a bit confused with
pick-axes when she want to use them as a weapon. since she often got
excal then a pick-axe is crap in comparison, and she ditch it. now she
doesn't consider it a weapon and won't ditch it