canidaeitized Dig
although, i broke this analyzer quite much in the process. for example,
it won't dig narrow passes when she stands next to one. also, i
significantly lowered the priority, and she won't start digging before
she's done exploring the level and is either just about to search
desperately or is heading for the downstairs.
on top of that, did not get time to actually test the changes