wooo! do take notice of this patch!
gimme a min, still giggling like a little schoolgirl..
...
ok, it'll be a bit long, but read it :)

i don't know why i didn't see this before, but while i was looking at
making dig widen narrow passages, i noticed that there's room for a
significant improvement in pathing!

consider this:
we're in the big room with 30 visible monsters, and we got no
projectiles. saiph will not call saiph->shortestPath(point) for each
monster to determine which one is closer. each time we call
saiph->shortestPath(point) we backtrack from the monster to saiph, which
means iterating a list of x nodes where x is amount of moves needed to
reach the monster. assume an average of 30 moves per monster and we're
iterating through 900 nodes each turn. sounds wasteful, doesn't it?

now consider this:
we know the cost for reaching each node (or tile, or square) on the map,
and the further away the monster is, the higher the cost will be (and
even higher if we need to cross traps/ice/altars/graves/fountains/etc).
instead of actually pathing to every monster we simply look at the cost
for reaching the tile the monster is standing on, then we path to the
monster with lowest path cost. voila! 900 iterations is now down to 60
(30 looking up path cost for each monster, 30 for pathing to nearest (as
in cost for pathing to) monster).

limitations:
- this can not be used on unreachable tiles, this means especially
  boulders and doors. saiph->shortestPath() can path to an unreachable
  tile, assuming one of the nearby squares are reachable.
- currently this can only be used on same level as we're standing on. it
  can be done for other levels too, but it requires some more logic, so
  i've delayed it for a while.
- as noted, this will cause saiph to go to the square with lowest path
  cost, not the square that requires least moves. however, 99% of the
  time this will be the exact same place, and 101% of the time, this is
  what you want her to do ;)
- as usual, this is not tested. i could've missed something and am
  making a complete ass out of myself. i intend to make
  saiph->shortestPath(unsigned char symbol) use this new "pathing" now,
  so the next commit should reveal whether i'm an ass or not :p