rewrote pathing
screw "getPathCost()", i came up with a better idea on my way home from
work (and skipped the gym to implement it).
saiph->shortestPath() methods now returns the new object PathNode which
contains the useful variables "dir", "moves" and "cost" which tells you
which direction to move to get to the point, amount of moves required to
get there, and the cost of getting there.

so, anything else that's new?
hell yeah!

saiph->shortestPath(Point) now runs in constant time. yeah, that's O(1).
nopes, i'm not shitting you. how? magic.
saiph->shortestPath(unsigned char) is quite close to constant time too,
but it needs to check other levels, so it's not just as fast (but still
bloody fast).
saiph->shortestPath(Coordinate) is slightly slower than that again, but
even this is ridiculously fast.