made variables in PathNode private
this is something i'd like to do different. even though we probably
won't need much information about a tile, it is an idea to make this a
class instead of just a character. further we can then inherit this tile
class into PathNode (or perhaps PathTile) which got the necessary extra
info for pathing