currently not compiling, temporary branch
working on tiles, and it requires some changes in the rest of the code,
most notably, the interlevel pathing requires some work. while i'm at it
i should improve this code at the same time. for one i should make a
cache telling whether to go up or from level a to get to level b (which
means each level must have a list of all other levels, but that's not a
problem, only about 70 or so unique levels anyways). should also improve
the code for updating path map, make another helper method, there's some
serious redundancy here. also, in theory pathing should be faster as it
won't clear the pathmap every time now, only update it (the cost(),
direction() and distance() returns UNREACHABLE if it wasn't updated).