Implement game data logging
saiph can now record the exact data of a previous game, and (provided
no changes have been made to AI behavior) retrace the same sequence
of internal states later.  This is expected to be helpful in debugging,
for instance allowing a weird situation to be rerun with custom debug
code.